bpf-ci-check: clarify message on existing PR
When the PR exists already, the script reuses it. Well, truth to tell,
it doesn't do much, because pushing to GitHub at the previous stage was
enough to update the PR and re-trigger a run of the workflow. But the
message "PR exists:" is ambiguous, it could be interpreted as an error
message indicating that nothing will happen. Let's clarify and state
that we "reuse" the existing PR.