Avoid overwriting bpf sources with bpf-next sources
The workflow file we download and commit to run selftests will download
the bpf-next sources if it's not executed on the kernel-patches/bpf
repository. This appears to be harmless if we are testing on bpf-next
(although it could be if we haven't rebased on latest bpf-next), but it
can easily cause issues if we are testing on bpf. In that last case, it
will overwrite all bpf sources with the bpf-next ones, potentially
leading to selftest failures such as:

    progs/test_btf_haskv.c:22:31: error: expected identifier
    BPF_ANNOTATE_KV_PAIR(btf_map, int, struct ipv_counts);
                                  ^
    progs/test_btf_haskv.c:22:1: error: type specifier missing, defaults to 'int'; ISO C99 and later do not support implicit int [-Wimplicit-int]
    BPF_ANNOTATE_KV_PAIR(btf_map, int, struct ipv_counts);
    ^
    int
    2 errors generated.

With this commit, we will edit the workflow on the fly to skip the
bpf-nxt download steps.

Signed-off-by: Paul Chaignon <paul@cilium.io>